requirements.txt
apache-airflow-providers-amazon==2.4.0
apache-airflow-providers-postgres==2.3.0
boto3==1.18.65
psycopg2-binary==2.9.2
dbt-core==1.1.1
dbt-redshift==1.1.0